+++ projects/slormancer-api/src/constants/common.ts
@@ -1,5 +1,6 @@
 export const MAX_HERO_LEVEL = 80;
 export const MAX_ITEM_LEVEL = MAX_HERO_LEVEL;
+export const MAX_ATTRIBUTE_RANK = 75;
 
 export const MAX_MAGIC_STATS = 1;
 export const MAX_RARE_STATS = 1;
@@ -38,8 +39,8 @@ export const POISON_DURATION = 7;
 export const ARCANE_BOND_DURATION = 7;
 export const TIME_LOCK_DURATION = 3;
 
-export const GAME_VERSION = '0.6.5l';
-export const API_VERSION = '0.4.0';
+export const GAME_VERSION = '0.6.5n';
+export const API_VERSION = '0.4.1';
 
 export const API_TO_GAME_VERSION_MAPPER: { [key: string]: string } = {
     '0.0.9': '0.3.015',
@@ -60,4 +61,5 @@ export const API_TO_GAME_VERSION_MAPPER: { [key: string]: string } = {
     '0.3.0': '0.5.01e',
     '0.3.1': '0.5.01e',
     '0.4.0': '0.6.5l',
+    '0.4.1': '0.6.5n',
 }