+++ CHANGELOG.md
@@ -28,6 +28,12 @@ Initial release
 - Fixed an issue causing ancestral legacy active effects to be applied twice
 - Fixed an issue causing exported build above level 64 to have an incorrect level
 
+## Release 0.4.1 (2023-25-07)
+## Fixes
+- Fixed an issue causing exported build with item level above 64 to have an incorrect level
+- Fixed an issue causing exported build with an attribute rank above 64 to have an incorrect rank
+- it is no longer possible to allocate more than 75 points into an attribute
+
 ## TODO
 - mise à jour packages
 - parsing loadouts